functions work, testing in place
