+++ test/compareNumbers.test.js
@@ -1,34 +1,38 @@
 // IMPORT MODULES under test here:
 // import example from '../src/example.js';
-import compareNumbers from '../compareNumbers'
+import compareNumbers from './compareNumbers.js'
 const test = QUnit.test;
 
-test('13 should return 0', function(assert) {
+test('guess was correct', function(assert) {
     //Arrange
     // Set up your parameters and expectations
     const guess = 13;
     const correctNumber = 13;
+    const expected = 0;
 
     //Act 
     // Call the function you're testing and set the result to a const
-    const result = compareNumbers(guess);
+    const result = compareNumbers(guess, correctNumber);
 
     //Assert
     // Make assertions about what is expected valid result
-    assert.equal(result, 0);
+    assert.equal(result, expected);
 });
 
-// We will need to know whether the user's guess is correct, too low, or too high.
+test('guess was too high', function(assert) {
+    const guess = 17;
+    const correctNumber = 13;
+    const expected = 1;
+    const result = compareNumbers(guess, correctNumber);
+    assert.equal(result, expected);
+});
 
-//TDD a compareNumbers function that:
-
-//Has two parameters: guess and correctNumber
-//Returns one of the following values:
-//0 - the numbers are identical
-//-1 - the guessed number is too low
-//1 - the guessed number is too high
-//Write one test at a time, each test should address one of the three possible outcomes.
-
-//You can assume valid input, meaning the passed input will always be a number.
+test('guess was too low', function(assert) {
+    const guess = 1;
+    const correctNumber = 13;
+    const expected = -1;
+    const result = compareNumbers(guess, correctNumber);
+    assert.equal(result, expected);
+});
 
 //SUPER STRETCH: Actually test for invalid input. You should throw an error if either parameter is not a number. Consult the QUnit docs for how to test for an exception.
